+++ cocos/core/components/ui-base/index.ts
@@ -27,9 +27,9 @@
  * @hidden
  */
 
-export * from './canvas-component';
+export * from './canvas';
 export * from './ui-component';
-export * from './ui-render-component';
-export * from './ui-transform-component';
+export * from './ui-renderable';
+export * from './ui-transform';
 
 import './deprecated';
