+++ cocos/ui/assembler/label/bmfont.ts
@@ -31,7 +31,7 @@ import { SpriteFrame } from '../../../core/assets/sprite-frame';
 import * as js from '../../../core/utils/js';
 import { Rect } from '../../../core/math';
 import { UI } from '../../../core/renderer/ui/ui';
-import { Label } from '../../components/label-component';
+import { Label } from '../../components/label';
 import { IAssembler } from '../../../core/renderer/ui/base';
 import { fillMeshVertices3D } from '../utils';
 import { bmfontUtils } from './bmfontUtils';
