+++ cocos/ui/assembler/label/ttf.ts
@@ -31,7 +31,7 @@
 import * as js from '../../../core/utils/js';
 import { Color, Vec3 } from '../../../core/math';
 import { UI } from '../../../core/renderer/ui/ui';
-import { Label } from '../../components/label-component';
+import { Label } from '../../components/label';
 import { IAssembler} from '../../../core/renderer/ui/base';
 import { ttfUtils } from './ttfUtils';
 import { IRenderData } from '../../../core/renderer/ui/render-data';
