+++ cocos/ui/components/editbox/edit-box-component.ts
@@ -37,8 +37,8 @@ import { Color, Size, Vec3 } from '../../../core/math';
 import { EventTouch } from '../../../core/platform';
 import { SystemEventType } from '../../../core/platform/event-manager/event-enum';
 import { Node } from '../../../core/scene-graph/node';
-import { Label, VerticalTextAlignment } from '../label-component';
-import { Sprite } from '../sprite-component';
+import { Label, VerticalTextAlignment } from '../label';
+import { Sprite } from '../sprite';
 import { EditBoxImpl } from './edit-box-impl';
 import { EditBoxImplBase } from './edit-box-impl-base';
 import { InputFlag, InputMode, KeyboardReturnType } from './types';
