+++ cocos/ui/components/progress-bar.ts
@@ -34,7 +34,7 @@ import { ccclass, help, executionOrder, menu, requireComponent, tooltip, type, r
 import { Size, Vec2, Vec3 } from '../../core/math';
 import { Enum } from '../../core/value-types';
 import { clamp01 } from '../../core/math/utils';
-import { Sprite } from './sprite-component';
+import { Sprite } from './sprite';
 import { warn } from '../../core/platform/debug';
 import { legacyCC } from '../../core/global-exports';
